LightRAG-TagSystem-Demo/requirements.txt
# LightRAG标签体系Demo依赖包 - DeepSeek版本
lightrag==0.1.0b6
openai==1.3.0
chromadb==0.4.15
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pandas==2.1.3
numpy==1.24.3
networkx==3.2
flask==3.0.0
pyyaml==6.0.1
requests==2.31.0
